7Modules/Jacob/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
aiosqlite==0.21.0
python-multipart==0.0.20
pydantic==2.11.7
python-jose[cryptography]==3.3.0
